Multiple changes, massive refactor:
- Move route, service, and utility things into their respective packages
- Add AbstractApplication to manage new mountable services
- Add HttpService that provides a simple interface for managing services, setting them up, and tearing them down
- Add AbstractService that constructs prefixes for custom routes to use (like in JsonService, previously "JsonApplication"), provides helper methods for instantiating an application
- Move routes into their respective services and mount them
- Simplify JsonService routes to automatically include the resource name as the prefix for the route, thereby making the route completely optional if desired
- Document many new classes and some old ones
- Move StaticFile inside the static file route
- Add Set of whitelisted routes and prefixes to skip authentication (for things like HTML/CSS/JS and the login page)
- Remove some println's
- Fix AbstractService test to implement new mountPoint method
