+++ src/main/java/com/johnstarich/moviematcher/controllers/HtmlService.java
@@ -1,18 +1,28 @@
 package com.johnstarich.moviematcher.controllers;
 
+import com.johnstarich.moviematcher.routes.ServeStaticFileRoute;
 import spark.Spark;
 
 /**
+ * Serves static HTML on /* with a default index.html served for files that cannot be found
  * Created by johnstarich on 5/22/16.
  */
-public class HtmlService extends BasicController {
+public class HtmlService implements HttpService {
 	@Override
-	public String prefix() {
+	public String mountPoint() {
 		return "/";
 	}
 
 	@Override
-	public void initService() {
-		Spark.get("/*", new ServeStaticFileRoute("/index.html"));
+	public String resource() {
+		return "";
 	}
+
+	@Override
+	public void init() {
+		Spark.get("/*", "text/html", new ServeStaticFileRoute("/index.html"));
+	}
+
+	@Override
+	public void destroy() {}
 }
