--- src/main/java/com/johnstarich/moviematcher/controllers/StaticFile.java
@@ -1,14 +1,0 @@
-package com.johnstarich.moviematcher.controllers;
-
-/**
- * Created by johnstarich on 4/7/16.
- */
-public class StaticFile {
-    public final String content;
-    public final String contentType;
-
-    public StaticFile(String content, String contentType) {
-        this.content = content;
-        this.contentType = contentType;
-    }
-}
