+++ src/main/java/com/johnstarich/moviematcher/utils/CountedSet.java
@@ -1,4 +1,4 @@
-package com.johnstarich.moviematcher;
+package com.johnstarich.moviematcher.utils;
 
 import java.util.Collection;
 import java.util.HashMap;
