+++ src/test/java/com/johnstarich/moviematcher/models/UserTest.java
@@ -1,6 +1,7 @@
 package com.johnstarich.moviematcher.models;
 
 import com.johnstarich.moviematcher.controllers.AbstractMongoDBTest;
+import com.johnstarich.moviematcher.utils.HttpException;
 import org.bson.types.ObjectId;
 
 import java.util.ArrayList;
@@ -50,7 +51,7 @@ public class UserTest extends AbstractMongoDBTest {
 
 		try {
 			John = John.resetPassword("wrongPassword", "INeedBetterPassword@2013");
-		} catch (com.johnstarich.moviematcher.app.HttpException e) {
+		} catch (HttpException e) {
 			assertEquals("Invalid old password.", e.getMessage());
 		}
